this just a madatory task

+++ 0x17-doubly_linked_lists/1-dlistint_len.c
@@ -7,7 +7,7 @@
 size_t dlistint_len(const dlistint_t *h)
 {
 	const dlistint_t *iterator;
-	size_t countnodes;
+	int countnodes;
 	countnodes = 0;
 
 	if (!h)
